Cleanup

+++ 2022/src/main/scala/aoc2022/day20/Solution.scala
@@ -6,17 +6,17 @@ import zio.stream.*
 
 object Solution {
 
-  def parseLine(line: String): Int = line.toInt
+  def parseLine(line: String): BigInt = BigInt(line.toInt)
 
-  def parseInput(lineStream: ZStream[Any, Throwable, String]): ZIO[Any, Throwable, List[Int]] = lineStream.map(parseLine).runCollect.map(_.toList)
+  def parseInput(lineStream: ZStream[Any, Throwable, String]): ZIO[Any, Throwable, List[BigInt]] = lineStream.map(parseLine).runCollect.map(_.toList)
 
-  def solvePart1(input: List[Int]): ZIO[Any, Throwable, Long] =
+  def solvePart1(input: List[BigInt]): ZIO[Any, Throwable, BigInt] =
     val decrypt = Decrypt(input)
     val output = decrypt.groveCoordinates(List(1000,2000,3000))
     ZIO.succeed(output.sum)
 
-  def solvePart2(input: List[Int]): ZIO[Any, Throwable, Long] =
-    val decrypt = Decrypt(input, 811589153L, 10)
+  def solvePart2(input: List[BigInt]): ZIO[Any, Throwable, BigInt] =
+    val decrypt = Decrypt(input, 811589153, 10)
     val output = decrypt.groveCoordinates(List(1000,2000,3000))
     ZIO.succeed(output.sum)
